Classes and interfaces associated with search filters split into separate files.
