+++ Analyzers/AssemblyPropertiesViewer.Analyzers.BasicAnalyzers/AssemblyFrameworkAnalyzer.cs
@@ -2,6 +2,7 @@
 using AssemblyPropertiesViewer.Analyzers.Interfaces;
 using AssemblyPropertiesViewer.Analyzers.Models;
 using AssemblyPropertiesViewer.Analyzers.Models.Filtering;
+using AssemblyPropertiesViewer.Analyzers.Models.Filtering.Interfaces;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Reflection;
