+++ Analyzers/Core/AssemblyPropertiesViewer.Analyzers.Loader/AssemblyProxy.cs
@@ -1,6 +1,6 @@
 ﻿using AssemblyPropertiesViewer.Analyzers.Interfaces;
 using AssemblyPropertiesViewer.Analyzers.Models;
-using AssemblyPropertiesViewer.Analyzers.Models.Filtering;
+using AssemblyPropertiesViewer.Analyzers.Models.Filtering.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.IO;
