+++ Clients/AssemblyPropertiesViewer.WPF/Services/Interfaces/IFilteringControlCreationService.cs
@@ -1,4 +1,4 @@
-﻿using AssemblyPropertiesViewer.Analyzers.Models.Filtering;
+﻿using AssemblyPropertiesViewer.Analyzers.Models.Filtering.Interfaces;
 using AssemblyPropertiesViewer.Controls;
 
 namespace AssemblyPropertiesViewer.Services.Interfaces
